added portainer to the installer

+++ installers/fedora.sh
@@ -130,7 +130,21 @@ rm ~/.config/scripts/wallpaper_file
 ln -s "$(pwd)/wallpapers/japanes-waves.svg" ~/.config/scripts/wallpaper_file
 
 # nyxt web browser
-curl -o nyxt.tar.xz https://nyxt.atlas.engineer/static/release/nyxt-linux-2.0.0.tar.xz
-sudo tar xf nyxt.tar.xz -C /
-rm nyxt.tar.xz
+if [ ! command -v nyxt &> /dev/null ]; then
+    curl -o nyxt.tar.xz https://nyxt.atlas.engineer/static/release/nyxt-linux-2.0.0.tar.xz
+    sudo tar xf nyxt.tar.xz -C /
+    rm nyxt.tar.xz
+fi
+rm -rf ~/.local/share/nyxt/extensions/nx-search-engines
 git clone https://github.com/aartaka/nx-search-engines ~/.local/share/nyxt/extensions/nx-search-engines
+
+# Portainer (Docker management)
+sudo docker top portainer || sudo docker run \
+    -d \
+    -p 8000:8000 \
+    -p 9000:9000 \
+    --name=portainer \
+    --restart=always \
+    -v /var/run/docker.sock:/var/run/docker.sock \
+    -v portainer_data:/data \
+    portainer/portainer-ce
